+++ src/main/java/org/sopt/dto/member/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package org.sopt.dto;
+package org.sopt.dto.member;
 
 
 public class MemberRequestDto {
