+++ src/main/java/org/sopt/dto/member/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package org.sopt.dto;
+package org.sopt.dto.member;
 
 public class MemberResponseDto {
     public static class Member{
